+++ sql/queries/feeds_follow.sql
@@ -8,6 +8,7 @@ RETURNING *;
 
 DELETE FROM follows_feeds
 WHERE id = $1
+AND user_id = $2
 RETURNING *;
 
 -- name: GetAllFeedFollowsByUser :many
